Making outbound handler protocol dependent via inbound handler
